% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmflow.R
\docType{data}
\name{europe2000}
\alias{europe2000}
\title{Predictive variables for Europe representing current times.}
\format{Raster brick with 22 layers.}
\usage{
europe2000
}
\description{
Raster brick with predictive variables to fit species distribution models for the European territory at ~20km resolution and reference system WGS84 (units are degrees of latitude and longitude, EPSG is 4326).
The variables are:
\itemize{
  \item \emph{bio1}: annual mean temperature (ºC x 10).
  \item \emph{bio2}: mean diurnal temperature range, computed as \eqn{mean of monthly (max temp - min temp)}.
  \item \emph{bio3}: isothermality, computed as \eqn{(bio2/bio7)*100}.
  \item \emph{bio4}: temperature seasonality (standard deviation of monthly averages * 100).
  \item \emph{bio5}: maximum temperature of the warmest month.
  \item \emph{bio6}: minimum temperature of the coldest month.
  \item \emph{bio7}: temperature annual range, computed as \eqn{(bio5-bio6}.
  \item \emph{bio8}: mean temperature of the wettest quarter.
  \item \emph{bio9}: mean temperature of the driest quarter.
  \item \emph{bio10}: mean temperature of the warmest quarter.
  \item \emph{bio11}: mean temperature of the coldest quarter.
  \item \emph{bio12}: annual precipitation (mm.).
  \item \emph{bio13}: precipitation of wettest month.
  \item \emph{bio14}: precipitation of driest month.
  \item \emph{bio15}: precipitation seasonality (standard deviation of monthly averages * 100).
  \item \emph{bio16}: precipitation of wettest quarter.
  \item \emph{bio17}: precipitation of driest quarter.
  \item \emph{bio18}: precipitation of warmest quarter.
  \item \emph{bio19}: precipitation of coldest quarter.
  \item \emph{ndvi_average}: average normalized vegetation index.
  \item \emph{ndvi_minimum}: minimum normalized vegetation index.
  \item \emph{ndvi_maximum}: maximum normalized vegetation index.
  \item \emph{topo_slope}: topographic slope (º).
}
}
\details{
Bioclim variables (\emph{bioXX}) were downloaded from the \emph{Worldclim v1.4} dataset (Hijmans et al. 2005, \url{https://www.worldclim.org/version1}).

The variable \emph{topo_slope} was generated from the same digital elevation model with the \emph{r.slope.aspect} function of the GRASS GIS software (Hofierkaet al. 2009).
}
\references{
Hijmans, R.J., S.E. Cameron, J.L. Parra, P.G. Jones and A. Jarvis, 2005. Very high resolution interpolated climate surfaces for global land areas. International Journal of Climatology 25: 1965-1978.

Sanderson, E.W., Jaiteh, M., Levy, M.A., Redford, K.H., Wannebo, A.V., Woolmer, G. (2002) The Human Footprint and the Last of the Wild: The human footprint is a global map of human influence on the land surface, which suggests that human beings are stewards of nature, whether we like it or not, BioScience, Volume 52, Issue 10, October 2002, Pages 891–904, https://doi.org/10.1641/0006-3568(2002)052[0891:THFATL]2.0.CO;2

Hofierka, J., Suri, M. (2002): The solar radiation model for Open source GIS: implementation and applications. International GRASS users conference in Trento, Italy, September 2002.

Jarvis, A., H.I. Reuter, A. Nelson, E. Guevara, 2008, Hole-filled SRTM for the globe Version 4, available from the CGIAR-CSI SRTM 90m Database (\url{http://srtm.csi.cgiar.org}).

Hofierka, J., Mitasova, H., Neteler, M., 2009. Geomorphometry in GRASS GIS. In: Hengl, T. and Reuter, H.I. (Eds), Geomorphometry: Concepts, Software, Applications. Developments in Soil Science, vol. 33, Elsevier, 387-410 pp, \url{http://www.geomorphometry.org}.
}
\keyword{datasets}
