% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_auto.R
\name{s_auto}
\alias{s_auto}
\title{Automatic selection of predictive variables for species distribution modeling}
\usage{
s_auto(
  training.df,
  response.col = "presence",
  select.cols = NULL,
  omit.cols = c("x", "y"),
  max.cor = 0.75,
  plot = TRUE,
  text.size = 6
)
}
\arguments{
\item{training.df}{A data frame with a presence column with 1 indicating presence and 0 indicating background, and columns with predictor values.}

\item{response.col}{Character, name of the presence column.}

\item{select.cols}{Character vector, names of the columns representing predictors. If \code{NULL}, all numeric variables but \code{response.col} are considered.}

\item{omit.cols}{Character vector, variables to exclude from the analysis.}

\item{max.cor}{Numeric in the interval [0, 1], maximum Pearson correlation of the selected variables. Defaults to 0.75.}

\item{plot}{Boolean, if \code{TRUE}, prints last correlation dendrogram to test the final output.}

\item{text.size}{Numeric, size of the dendrogram labels.}
}
\value{
An object of class \code{s_auto} with three named slots. The slot \code{plot} contains a correlation dendrogram produced by \code{\link{s_cor}}. The slot \code{df} contains the VIF data frame of the selected variables produced by \code{\link{s_vif_auto}}. The slot \code{vars} is a character vector with the names of the selected variables.
}
\description{
Applies \code{\link{s_biserial_cor}}, \code{\link{s_lower_vif}}, and \code{\link{s_lower_cor}} to automatically select a set of non-correlated variables with a biserial correlation as high as possible.
}
\examples{
\dontrun{
data(virtual.species.training)
selected.vars <- s_auto(
  training.df = virtual.species.training,
  response.col = "presence",
  omit.cols = c("x", "y")
)
selected.vars

HH::vif(virtual.species.training[, selected.vars])
cor(virtual.species.training[, selected.vars])

}

}
\author{
Blas Benito <blasbenito@gmail.com>.
}
