% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_biserial_cor.R
\name{s_biserial_cor}
\alias{s_biserial_cor}
\title{Biserial correlation analysis of presence and background data for variable selection}
\usage{
s_biserial_cor(
  training.df,
  response.col = "presence",
  select.cols = NULL,
  omit.cols = c("x", "y"),
  axis.text.size = 6,
  legend.text.size = 12,
  strip.text.size = 10,
  point.size = 1,
  line.size = 1,
  plot = TRUE
)
}
\arguments{
\item{training.df}{A data frame with a presence column with 1 indicating presence and 0 indicating background, and columns with predictor values.}

\item{response.col}{Character, name of the presence column.}

\item{select.cols}{Character vector, names of the columns representing predictors. If \code{NULL}, all numeric variables but \code{response.col} are considered.}

\item{omit.cols}{Character vector, variables to exclude from the analysis.}

\item{axis.text.size}{Numeric, size of the axis labels.}

\item{legend.text.size}{Numeric, size of the legend labels.}

\item{strip.text.size}{Numeric, size of the panel names.}

\item{point.size}{Size of points in the biserial correlation plot.}

\item{line.size}{Line width in the biserial correlation plot.}

\item{plot}{Boolean, prints biserial correlation plot if \code{TRUE}.}
}
\value{
A named list with two slots named \code{plot} and \code{df}. The former contains a ggplot object with the biserial correlation analysis. The latter is a data frame with the following columns:
\itemize{
  \emph{variable}: Name of the predictive variable.
  \emph{R2}: R-squared of the biserial correlation.
  \emph{p}: p-value of the correlation analysis.
}
The output data frame is ordered, starting with the higher R2 values.
}
\description{
Computes the biserial correlation between presence and background data for a set of predictors. A high biserial correlation for a given predictor indicates that the distributions of the presence and background records are separated enough in the space of predictor values to suggest that the predictor is a good candidate for a species distribution model.
}
\examples{
data(virtualSpeciesPB)
cPB <- s_biserial_cor(
  training.df = virtualSpeciesPB,
  response.col = "presence",
  select.cols = c("bio1", "bio5", "bio6")
)

}
\author{
Blas Benito <blasbenito@gmail.com>
}
