% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_cor.R
\name{s_cor}
\alias{s_cor}
\title{Correlation dendrogram to help reduce multicollinearity in a training dataset.}
\usage{
s_cor(
  training.df,
  select.cols = NULL,
  omit.cols = c("x", "y", "presence"),
  max.cor = 0.5,
  biserial.cor = NULL,
  plot = TRUE,
  text.size = 6
  )
}
\arguments{
\item{training.df}{A data frame with a presence column with 1 indicating presence and 0 indicating background, and columns with predictor values.}

\item{select.cols}{Character vector, names of the columns representing predictors. If \code{NULL}, all numeric variables but \code{presence.column} are considered.}

\item{omit.cols}{Character vector, variables to exclude from the analysis. Defaults to \code{c("x", "y", "presence")}.}

\item{biserial.cor}{List, output of the function \code{\link{s_biserial_cor}}. Its R-squared scores are used to select variables.}

\item{plot}{Boolean, prints biserial correlation plot if \code{TRUE}.}

\item{text.size}{Numeric, size of the dendrogram labels.}

\item{maximum.cor}{Numeric in the interval [0, 1], maximum Pearson correlation of the selected variables.}
}
\value{
If \code{biserial.cor} is not NULL, a list with two slots named "dendrogram" (a ggplot2 object) and "selected.variables" with the dendrogram and the character vector with the selected variables. Otherwise it only returns the dendrogram, and the users have to select the variables by themselves.
}
\description{
Computes the correlation between all pairs of variables in a training dataset. If a \code{\link{s_biserial_cor}} output is provided, it further selects variables automatically based on the R-squared value obtained by each variable in the biserial correlation analysis.
}
\examples{
\dontrun{
data("virtualSpeciesPB")

biserial.cor <- s_biscor(
 training.df = virtualSpeciesPB,
 omit.cols = c("x", "y")
)

selected.vars <- s_cor(
 training.df = virtualSpeciesPB,
 select.cols = NULL,
 omit.cols = c("x", "y", "presence"),
 max.cor = 0.5,
 biserial.cor = biserial.cor
)$selected.variables
}

}
\author{
Blas Benito <blasbenito@gmail.com>.
}
