% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_cor_auto.R
\name{s_cor_auto}
\alias{s_cor_auto}
\title{Reduces multicollinearity automatically through repeated application of \code{\link{s_cor}}.}
\usage{
s_cor_auto(
training.df,
select.cols = NULL,
omit.cols = c("x", "y", "presence"),
max.cor = 0.5,
biserial.cor = NULL,
plot = TRUE,
text.size = 6
)
}
\arguments{
\item{training.df}{A data frame with a presence column with 1 indicating presence and 0 indicating background, and columns with predictor values.}

\item{select.cols}{Character vector, names of the columns representing predictors. If \code{NULL}, all numeric variables but \code{presence.column} are considered.}

\item{omit.cols}{Character vector, variables to exclude from the analysis. Defaults to \code{c("x", "y", "presence")}.}

\item{max.cor}{Numeric in the interval [0, 1], maximum Pearson correlation of the selected variables. Defaults to 0.5.}

\item{biserial.cor}{List, output of the function \code{\link{s_biserial_cor}}. Its R-squared scores are used to select variables.}

\item{plot}{Boolean, prints biserial correlation plot if \code{TRUE}.}

\item{text.size}{Numeric, size of the dendrogram labels.}
}
\value{
A character vector with the names of the selected variables.
}
\description{
Computes the correlation between all pairs of variables in a training dataset and removes variables until there are no more variables correlated above a given threshold (0.5 Pearson correlation by default). This function requires a \code{\link{s_biserial_cor}} output, and it won't run if the argument remains \code{NULL}.
}
\examples{
\dontrun{
data("virtualSpeciesPB")

biserial.cor <- s_biserial_cor(
 training.df = virtualSpeciesPB,
 omit.cols = c("x", "y")
)

selected.vars <- s_cor_auto(
 training.df = virtualSpeciesPB,
 omit.cols = c("x", "y", "presence"),
 max.cor = 0.5,
 biserial.cor = biserial.cor
)$selected.variables
}

}
\author{
Blas Benito <blasbenito@gmail.com>.
}
