% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_plot_density.R
\name{s_plot_density}
\alias{s_plot_density}
\title{Plots density functions of use and availability from a training data frame}
\usage{
s_plot_density(
  training.df,
  response.col = "presence",
  select.cols = NULL,
  omit.cols = c("x", "y"),
  axis.text.size = 6,
  legend.text.size = 12,
  strip.text.size = 10
)
}
\arguments{
\item{training.df}{A data frame with a presence column with 1 indicating presence and 0 indicating background, and columns with predictor values.}

\item{response.col}{Character, name of the response variable. Usually a presence-background column with ones and zeroes.}

\item{select.cols}{Character vector, names of the columns representing predictors. If \code{NULL}, all numeric variables but \code{response.col} are considered.}

\item{omit.cols}{Character vector, variables to exclude from the analysis.}

\item{axis.text.size}{Numeric, size of the axis labels.}

\item{legend.text.size}{Numeric, size of the legend labels.}

\item{strip.text.size}{Numeric, size of the panel names.}
}
\value{
A ggplot object.
}
\description{
Plots the relative densities of ones (presence, a.k.a, "use") versus zeroes (background, a.k.a, "availability") for each environmental predictor in a training data frame. This plot helps to understand the relationship between use and availability in order to make informed decisions during variable selection. When for a given variable the density of use peaks over low availability it indicates that the species selects those values of a variable at a higher rate than what is expected by chance. On the other hand, variables with a very high overlap between use and availability will likely turn out to have a low predictive value during SDM fitting.
}
\examples{
data("virtual.species.training")
x <- s_plot_density(
 x = virtual.species.training,
 response.col = "presence",
 select.cols = NULL,
 omit.cols = c("x", "y")
)

}
\author{
Blas Benito <blasbenito@gmail.com>
}
