% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_auto.R
\name{s_auto}
\alias{s_auto}
\title{Automatic selection of predictive variables for species distribution modeling}
\usage{
s_auto(
  training.df,
  response.col = "presence",
  select.cols = NULL,
  omit.cols = NULL,
  plot = TRUE,
  text.size = 6
)
}
\arguments{
\item{training.df}{A data frame with a presence column with 1 indicating presence and 0 indicating background, and columns with predictor values.}

\item{response.col}{Character, name of the presence column.}

\item{select.cols}{Character vector, names of the columns representing predictors. If \code{NULL}, all numeric variables but \code{response.col} are considered.}

\item{omit.cols}{Character vector, variables to exclude from the analysis.}

\item{plot}{Boolean, if \code{TRUE}, prints last correlation dendrogram to test the final output.}

\item{text.size}{Numeric, size of the dendrogram labels.}
}
\value{
A character vector with the names of the selected variables.
}
\description{
Applies \code{\link{s_biserial_cor}}, \code{\link{s_cor}} (with correlation threshold set to 0.5), and \code{\link{s_auto_vif}} to automatically select a set of non-correlated variables with the higher biserial correlation as possible.
}
\examples{
\dontrun{
data(virtualSpeciesPB)
selected.vars <- s_auto(
  training.df = virtualSpeciesPB,
  response.col = "presence",
  omit.cols = c("x", "y")
)
selected.vars
HH::vif(virtualSpeciesPB[, selected.vars])
cor(virtualSpeciesPB[, selected.vars])
}

}
\author{
Blas Benito <blasbenito@gmail.com>.
}
